+++ tests/testthat/test_scale.R
@@ -27,3 +27,31 @@ expect_equal(min(iris10_range), 0)
 iris10_colrange <- scale_input(iris10, scale_type = "colrange")
 expect_equal(apply(iris10_colrange, 2, max), rep(1, 4), check.attributes = FALSE)
 expect_equal(apply(iris10_colrange, 2, min), rep(0, 4), check.attributes = FALSE)
+
+test_that("scaling applied outside umap should not appear in the model", {
+  iris10s <- scale(iris10)
+  iris10s_umap0 <-
+    umap(
+      iris10s,
+      n_neighbors = 4,
+      n_epochs = 0,
+      init = "rand",
+      ret_model = TRUE,
+      scale = FALSE
+    )
+  expect_null(iris10s_umap0$scale_info)
+
+  iris10s_umap0s <-
+    umap(
+      iris10s,
+      n_neighbors = 4,
+      n_epochs = 0,
+      init = "rand",
+      ret_model = TRUE,
+      scale = TRUE
+    )
+  expect_equal(
+    names(iris10s_umap0s$scale_info),
+    c("scaled:center", "scaled:scale", "scaled:nzvcols")
+  )
+})